+++ app/src/main/java/org/nanolx/securitybypasser/XposedEntryPoint.java
@@ -6,6 +6,7 @@ import org.nanolx.securitybypasser.apps.AmazonFlex;
 import org.nanolx.securitybypasser.apps.HIT;
 import org.nanolx.securitybypasser.apps.KingdomHeartsUX;
 import org.nanolx.securitybypasser.apps.Miitomo;
+import org.nanolx.securitybypasser.apps.MonsterHunterExplore;
 import org.nanolx.securitybypasser.apps.PuzzlesAndDragons;
 import org.nanolx.securitybypasser.apps.YokaiWatchWW;
 
@@ -27,6 +28,7 @@ public class XposedEntryPoint implements IXposedHookLoadPackage
 		appRegistry.put("com.nexon.hit.global", HIT.class);
 		appRegistry.put("com.square_enix.android_googleplay.khuxww", KingdomHeartsUX.class);
 		appRegistry.put("com.nintendo.zaaa", Miitomo.class);
+		appRegistry.put("jp.co.capcom.android.explore", MonsterHunterExplore.class);
 		appRegistry.put("jp.gungho.padEN", PuzzlesAndDragons.class);
 		appRegistry.put("com.level5.ywwwus", YokaiWatchWW.class);
 	}